Renamed package 'default' to 'standard' to prevent conflicts with Java keyword 'default'
