+++ src/main/kotlin/com/github/magneticflux/rss/Persisters.kt
@@ -1,25 +1,5 @@
 package com.github.magneticflux.rss
 
-import com.github.magneticflux.rss.namespaces.default.converters.CategoryConverter
-import com.github.magneticflux.rss.namespaces.default.converters.ChannelConverter
-import com.github.magneticflux.rss.namespaces.default.converters.CloudConverter
-import com.github.magneticflux.rss.namespaces.default.converters.EnclosureConverter
-import com.github.magneticflux.rss.namespaces.default.converters.GuidConverter
-import com.github.magneticflux.rss.namespaces.default.converters.ImageConverter
-import com.github.magneticflux.rss.namespaces.default.converters.ItemConverter
-import com.github.magneticflux.rss.namespaces.default.converters.RssFeedConverter
-import com.github.magneticflux.rss.namespaces.default.converters.SourceConverter
-import com.github.magneticflux.rss.namespaces.default.converters.TextInputConverter
-import com.github.magneticflux.rss.namespaces.default.elements.Category
-import com.github.magneticflux.rss.namespaces.default.elements.Channel
-import com.github.magneticflux.rss.namespaces.default.elements.Cloud
-import com.github.magneticflux.rss.namespaces.default.elements.Enclosure
-import com.github.magneticflux.rss.namespaces.default.elements.Guid
-import com.github.magneticflux.rss.namespaces.default.elements.Image
-import com.github.magneticflux.rss.namespaces.default.elements.Item
-import com.github.magneticflux.rss.namespaces.default.elements.RssFeed
-import com.github.magneticflux.rss.namespaces.default.elements.Source
-import com.github.magneticflux.rss.namespaces.default.elements.TextInput
 import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesBlockConverter
 import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesCompleteConverter
 import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesDurationConverter
@@ -32,6 +12,26 @@ import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesDuration
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesImage
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesSubCategory
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesTopLevelCategory
+import com.github.magneticflux.rss.namespaces.standard.converters.CategoryConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.ChannelConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.CloudConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.EnclosureConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.GuidConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.ImageConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.ItemConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.RssFeedConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.SourceConverter
+import com.github.magneticflux.rss.namespaces.standard.converters.TextInputConverter
+import com.github.magneticflux.rss.namespaces.standard.elements.Category
+import com.github.magneticflux.rss.namespaces.standard.elements.Channel
+import com.github.magneticflux.rss.namespaces.standard.elements.Cloud
+import com.github.magneticflux.rss.namespaces.standard.elements.Enclosure
+import com.github.magneticflux.rss.namespaces.standard.elements.Guid
+import com.github.magneticflux.rss.namespaces.standard.elements.Image
+import com.github.magneticflux.rss.namespaces.standard.elements.Item
+import com.github.magneticflux.rss.namespaces.standard.elements.RssFeed
+import com.github.magneticflux.rss.namespaces.standard.elements.Source
+import com.github.magneticflux.rss.namespaces.standard.elements.TextInput
 import org.simpleframework.xml.convert.Registry
 import org.simpleframework.xml.convert.RegistryStrategy
 import org.simpleframework.xml.core.Persister
