+++ src/main/kotlin/com/github/magneticflux/rss/namespaces/standard/converters/SourceConverter.kt
@@ -1,7 +1,7 @@
-package com.github.magneticflux.rss.namespaces.default.converters
+package com.github.magneticflux.rss.namespaces.standard.converters
 
 import com.github.magneticflux.rss.URLTransform
-import com.github.magneticflux.rss.namespaces.default.elements.Source
+import com.github.magneticflux.rss.namespaces.standard.elements.Source
 import org.simpleframework.xml.convert.Converter
 import org.simpleframework.xml.stream.InputNode
 import org.simpleframework.xml.stream.OutputNode
