+++ src/main/kotlin/com/github/magneticflux/rss/namespaces/standard/elements/Channel.kt
@@ -1,10 +1,10 @@
-package com.github.magneticflux.rss.namespaces.default.elements
+package com.github.magneticflux.rss.namespaces.standard.elements
 
-import com.github.magneticflux.rss.namespaces.default.converters.ChannelConverter
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesBlock
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesExplicitStatus
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesImage
 import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesTopLevelCategory
+import com.github.magneticflux.rss.namespaces.standard.converters.ChannelConverter
 import org.simpleframework.xml.Root
 import org.threeten.bp.DayOfWeek
 import org.threeten.bp.ZonedDateTime
